packages/core/pyproject.toml
[project]
name = "core"
version = "0.0.0"
description = "Shared domain logic for the voice AI template"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "httpx>=0.27",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    "structlog>=24.4",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "types-python-jose>=3.3.4.20240106",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
markers = [
    "integration: marks tests as integration (deselect with '-m \"not integration\"')",
]
